+++ src/tqec/compile/__init__.py
@@ -11,6 +11,7 @@ on available hardware.
 from .block import CompiledBlock as CompiledBlock
 from .compile import CompiledGraph as CompiledGraph
 from .compile import compile_block_graph as compile_block_graph
+from .generation import generate_circuit as generate_circuit
 from .specs import BlockBuilder as BlockBuilder
 from .specs import CubeSpec as CubeSpec
 from .specs import PipeSpec as PipeSpec
