+++ src/tqec/compile/block.py
@@ -3,13 +3,13 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Mapping
 
-from tqec.circuit.generation import generate_circuit
 from tqec.circuit.schedule import ScheduledCircuit
-from tqec.utils.exceptions import TQECException
+from tqec.compile.generation import generate_circuit
 from tqec.plaquette.library.empty import empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquette, Plaquettes, RepeatedPlaquettes
-from tqec.templates.indices.base import RectangularTemplate
-from tqec.templates.indices.layout import LayoutTemplate
+from tqec.templates.base import RectangularTemplate
+from tqec.templates.layout import LayoutTemplate
+from tqec.utils.exceptions import TQECException
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
 from tqec.utils.position import BlockPosition2D, Shift2D
 from tqec.utils.scale import LinearFunction
@@ -161,7 +161,7 @@ class BlockLayout:
             as many circuits as there are layers in ``self``, each circuit being
             the instantiation of one layer (i.e., a set of
             :class:`~tqec.plaquette.plaquette.Plaquette` instances) and the
-            :class:`~tqec.templates.indices.base.Template` instance from ``self``.
+            :class:`~tqec.templates.base.Template` instance from ``self``.
         """
         # We need to shift the circuit based on the shift of the layout template.
         top_left_plaquette = self._template.instantiation_origin(k)
