+++ src/tqec/compile/detectors/compute_test.py
@@ -25,10 +25,10 @@ from tqec.utils.exceptions import TQECException
 from tqec.plaquette.library.css import make_css_surface_code_plaquette
 from tqec.plaquette.library.empty import empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquettes
-from tqec.templates.indices._testing import FixedTemplate
-from tqec.templates.indices.layout import LayoutTemplate
-from tqec.templates.indices.qubit import QubitTemplate
-from tqec.templates.indices.subtemplates import SubTemplateType
+from tqec.templates._testing import FixedTemplate
+from tqec.templates.layout import LayoutTemplate
+from tqec.templates.qubit import QubitTemplate
+from tqec.templates.subtemplates import SubTemplateType
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
 from tqec.utils.position import BlockPosition2D, Shift2D
