+++ src/tqec/compile/specs/library/_utils.py
@@ -7,7 +7,7 @@ from tqec.utils.enums import Basis
 from tqec.plaquette.enums import PlaquetteOrientation, PlaquetteSide
 from tqec.plaquette.library import PlaquetteBuilder, empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquette, Plaquettes
-from tqec.templates.indices.qubit import QubitTemplate
+from tqec.templates.qubit import QubitTemplate
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
 from tqec.utils.position import Direction3D
 from tqec.utils.scale import LinearFunction
