+++ src/tqec/plaquette/rpng/rpng_test.py
