+++ src/tqec/plaquette/rpng/translators/base.py
@@ -1,7 +1,6 @@
 from abc import ABC, abstractmethod
 
 from tqec.plaquette.plaquette import Plaquette
-from tqec.plaquette.qubit import PlaquetteQubits, SquarePlaquetteQubits
 from tqec.plaquette.rpng import RPNGDescription
 
 
@@ -15,14 +14,12 @@ class RPNGTranslator(ABC):
     def translate(
         self,
         rpng_description: RPNGDescription,
-        qubits: PlaquetteQubits = SquarePlaquetteQubits(),
     ) -> Plaquette:
         """Generate the plaquette corresponding to the provided ``RPNG``
         description.
 
         Args:
             rpng_description: description of the plaquette to generate.
-            qubits: qubits to use for the returned plaquette.
 
         Returns:
             a valid implementation of the provided
