+++ src/tqec/plaquette/rpng/visualisation.py
