+++ src/tqec/templates/enums.py
@@ -1,3 +1,5 @@
+"""Defines enumerations that are only used in :mod:`tqec.templates`."""
+
 from enum import Enum, auto
 from typing import Literal
 
@@ -11,3 +13,10 @@ class ZObservableOrientation(Enum):
 
     def vertical_basis(self) -> Literal["x", "z"]:
         return "z" if self == ZObservableOrientation.VERTICAL else "x"
+
+
+class TemplateBorder(Enum):
+    TOP = auto()
+    BOTTOM = auto()
+    LEFT = auto()
+    RIGHT = auto()
