--- src/tqec/templates/indices/enums.py
@@ -1,8 +1,0 @@
-from enum import Enum, auto
-
-
-class TemplateBorder(Enum):
-    TOP = auto()
-    BOTTOM = auto()
-    LEFT = auto()
-    RIGHT = auto()
