+++ src/tqec/templates/qubit.py
@@ -8,8 +8,8 @@ import numpy.typing as npt
 from typing_extensions import override
 
 from tqec.utils.exceptions import TQECException, TQECWarning
-from tqec.templates.indices.base import BorderIndices, RectangularTemplate
-from tqec.templates.indices.enums import TemplateBorder
+from tqec.templates.base import BorderIndices, RectangularTemplate
+from tqec.templates.enums import TemplateBorder
 from tqec.utils.scale import LinearFunction, Scalable2D
