+++ src/tqec/templates/qubit_test.py
@@ -2,8 +2,8 @@ import numpy
 import pytest
 
 from tqec.utils.exceptions import TQECException, TQECWarning
-from tqec.templates.indices.enums import TemplateBorder
-from tqec.templates.indices.qubit import (
+from tqec.templates.enums import TemplateBorder
+from tqec.templates.qubit import (
     QubitHorizontalBorders,
     QubitSpatialCubeTemplate,
     QubitTemplate,
